Formatando valores finais como inteiros

+++ main.py
@@ -17,20 +17,20 @@ def obter_input_int(mensagem):
             print("Por favor, insira um número inteiro válido.")
 
 # Obter os valores de entrada
-Pu = obter_input_int("Insira o valor da Potência Útil (em W): ")
+Pu = int(obter_input_int("Insira o valor da Potência Útil (em W): "))
 
-n = obter_input_float("Insira o valor do Rendimento (em %): ") / 100
+n = float(obter_input_float("Insira o valor do Rendimento (em %): ") / 100)
 
-fp = obter_input_float("Insira o valor do Fator Potência (0 a 1): ")
+fp = float(obter_input_float("Insira o valor do Fator Potência (0 a 1): "))
 
 # (P) = Cálculo da Potência Ativa (Em W.)
-P = Pu / n
+P = int(Pu / n)
 
 # (S) = Cálculo da Potência Aparente (Em VA.)
-S = P / fp
+S = int(P / fp)
 
 # (Q) = Cálculo da Potência Reativa (Em VAr.)
-Q = math.sqrt(S**2 - P**2)
+Q = int(math.sqrt(S**2 - P**2))
 
 # Exibir os resultados
 print("-----------------------------------------")
